% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traceline.r
\name{categorical_traceline_pts}
\alias{categorical_traceline_pts}
\title{Ordinal item tracelines (graded response).}
\usage{
categorical_traceline_pts(
  p_active,
  theta,
  predictor.data,
  samp_size,
  num_responses_item,
  num_quadpts
)
}
\arguments{
\item{p_active}{Vector of item parameters.}

\item{theta}{Matrix of adaptive theta values.}

\item{predictor.data}{Matrix or dataframe of DIF and/or impact predictors.}

\item{samp_size}{Sample size in dataset.}

\item{num_responses_item}{Number of responses for item.}

\item{num_quadpts}{Number of quadrature points used for approximating the
latent variable.}
}
\description{
Ordinal item tracelines (graded response).
}
\keyword{internal}
