% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traceline.r
\name{cumulative_traceline_pts}
\alias{cumulative_traceline_pts}
\title{Ordinal tracelines (for derivatives).}
\usage{
cumulative_traceline_pts(
  p_active,
  theta,
  pred.data,
  samp_size,
  num_responses_item,
  quad.pts
)
}
\arguments{
\item{p_active}{Vector of item parameters.}

\item{theta}{Matrix of adaptive theta values.}

\item{pred.data}{Matrix or dataframe of DIF and/or impact predictors.}

\item{samp_size}{Sample size in dataset.}

\item{num_responses_item}{Number of responses for item.}

\item{quad.pts}{Number of quadrature points used for approximating the
latent variable.}
}
\description{
Ordinal tracelines (for derivatives).
}
\keyword{internal}
