% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traceline.r
\name{cumulative_traceline_pts_proxy}
\alias{cumulative_traceline_pts_proxy}
\title{Ordinal tracelines using proxy data.}
\usage{
cumulative_traceline_pts_proxy(
  p_item,
  prox_data,
  pred_data,
  samp_size,
  num_responses_item
)
}
\arguments{
\item{p_item}{Vector of item parameters.}

\item{prox_data}{Vector of observed proxy scores.}

\item{pred_data}{Matrix or dataframe of DIF and/or impact predictors.}

\item{samp_size}{Sample size in dataset.}

\item{num_responses_item}{Number of responses for item.}
}
\value{
a \code{"matrix"} of probability values for categorical (cumulative) item likelihood
}
\description{
Ordinal tracelines using proxy data.
}
\keyword{internal}
