% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.regDIF.R
\name{cv.regDIF}
\alias{cv.regDIF}
\title{Cross-Validation for regDIF}
\usage{
cv.regDIF(item.data,
       predictor.data,
       ...,
       nfolds = 10,
       fold,
       seed)
}
\arguments{
\item{item.data}{Matrix or dataframe of item responses, as in regDIF.}

\item{predictor.data}{Matrix or dataframe of DIF and/or impact predictors, as
in regDIF.}

\item{...}{Additional arguments to regDIF.}

\item{nfolds}{Numeric value indicating the number of cross-validation folds.
The default is 10.}

\item{fold}{The fold that each observation belongs to. The default is to
assign the observations randomly.}

\item{seed}{Numeric value representing the seed for the random number
generator. Used to obtain reproducible results.}
}
\value{
Function returns an object of class \code{cv.regDIF}
}
\description{
Use k-fold cross validation to identify best-fitting model over a grid of
tau values (i.e., regularization tuning parameter).
}
\examples{
\dontrun{

library(regDIF)
head(ida)
item.data <- ida[,1:6]
predictor.data <- ida[,7:9]
cv.fit <- cv.regDIF(item.data, predictor.data)
summary(cv.fit)

}

}
