% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derivatives.r
\name{d_alpha}
\alias{d_alpha}
\title{Partial derivatives for mean impact equation.}
\usage{
d_alpha(
  p_impact,
  etable,
  theta,
  mean_predictors,
  var_predictors,
  cov,
  samp_size,
  num_items,
  num.quad
)
}
\arguments{
\item{theta}{Matrix of adaptive theta values.}

\item{mean_predictors}{Possibly different matrix of predictors for the mean
impact equation.}

\item{var_predictors}{Possibly different matrix of predictors for the
variance impact equation.}

\item{cov}{Covariate being maximized.}

\item{samp_size}{Sample size in data set.}

\item{num_items}{Number of items in data set.}

\item{num.quad}{Number of quadrature points used for approximating the
latent variable.}

\item{p_active}{Vector of item parameters.}

\item{etable_all}{E-table for impact.}
}
\description{
Partial derivatives for mean impact equation.
}
\keyword{internal}
