% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derivatives.r
\name{d_bernoulli_itemblock_proxy}
\alias{d_bernoulli_itemblock_proxy}
\title{Partial derivatives for binary items by item-blocks using observed score proxy.}
\usage{
d_bernoulli_itemblock_proxy(
  p_item,
  pred_data,
  item_data_current,
  prox_data,
  samp_size,
  num_items,
  num_predictors,
  num_quad
)
}
\arguments{
\item{p_item}{Vector of item parameters.}

\item{pred_data}{Matrix or dataframe of DIF and/or impact predictors.}

\item{item_data_current}{Vector of current item responses.}

\item{prox_data}{Vector of observed proxy scores.}

\item{samp_size}{Sample size in dataset.}

\item{num_items}{Number of items in dataset.}

\item{num_predictors}{Number of predictors in dataset.}
}
\value{
a \code{"list"} of first and second partial derivatives for Bernoulli item likelihood (to
use with multivariate Newton-Raphson and observed proxy scores)
}
\description{
Partial derivatives for binary items by item-blocks using observed score proxy.
}
\keyword{internal}
