% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derivatives.r
\name{d_categorical_itemblock}
\alias{d_categorical_itemblock}
\title{Partial derivatives for ordinal items.}
\usage{
d_categorical_itemblock(
  parm,
  p_item,
  etable_item,
  theta,
  pred_data,
  thr,
  cov,
  samp_size,
  num_responses_item,
  num_items,
  num_quad
)
}
\arguments{
\item{parm}{Item parameter being maximized.}

\item{p_item}{Vector of item parameters.}

\item{etable_item}{E-table for impact.}

\item{theta}{Matrix of adaptive theta values.}

\item{thr}{Threshold value being maximized.}

\item{cov}{Covariate being maximized.}

\item{samp_size}{Sample size in dataset.}

\item{num_items}{Number of items in dataset.}

\item{pred.data}{Matrix or dataframe of DIF and/or impact predictors.}

\item{num.quad}{Number of quadrature points used for approximating the
latent variable.}
}
\description{
Partial derivatives for ordinal items.
}
\keyword{internal}
