% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derivatives.r
\name{d_categorical_proxy}
\alias{d_categorical_proxy}
\title{Partial derivatives for ordinal items using proxy data.}
\usage{
d_categorical_proxy(
  parm,
  p_item,
  prox_data,
  pred_data,
  item_data_current,
  thr,
  cov,
  samp_size,
  num_responses_item,
  num_items
)
}
\arguments{
\item{parm}{Item parameter being maximized.}

\item{p_item}{Vector of item parameters.}

\item{prox_data}{Vector of observed proxy scores.}

\item{pred_data}{Matrix or dataframe of DIF and/or impact predictors.}

\item{item_data_current}{Vector of current item responses.}

\item{thr}{Threshold value being maximized.}

\item{cov}{Covariate being maximized.}

\item{samp_size}{Sample size in dataset.}

\item{num_responses_item}{Number of responses for item.}

\item{num_items}{Number of items in dataset.}
}
\value{
a \code{"list"} of first and second partial derivatives for categorical item likelihood
(to use with coordinate descent and univariate Newton-Raphson)
}
\description{
Partial derivatives for ordinal items using proxy data.
}
\keyword{internal}
