% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derivatives.r
\name{d_mu_gaussian_proxy}
\alias{d_mu_gaussian_proxy}
\title{Partial derivatives for mean parameter of continuous items with proxy data.}
\usage{
d_mu_gaussian_proxy(
  parm,
  p_item,
  prox_data,
  responses_item,
  pred_data,
  cov,
  samp_size
)
}
\arguments{
\item{parm}{Item parameter being maximized.}

\item{p_item}{Vector of item parameters.}

\item{prox_data}{Vector of observed proxy scores.}

\item{responses_item}{Vector of item responses.}

\item{pred_data}{Matrix or dataframe of DIF and/or impact predictors.}

\item{cov}{Covariate being maximized.}

\item{samp_size}{Sample size in dataset.}
}
\value{
a \code{"list"} of first and second partial derivatives for mean value of Gaussian item
likelihood (to use with coordinate descent and univariate Newton-Raphson)
}
\description{
Partial derivatives for mean parameter of continuous items with proxy data.
}
\keyword{internal}
