% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derivatives.r
\name{d_sigma_gaussian}
\alias{d_sigma_gaussian}
\title{Partial derivatives for variance parameter of continuous items.}
\usage{
d_sigma_gaussian(
  parm,
  p_item,
  etable,
  theta,
  responses_item,
  pred.data,
  cov,
  samp_size,
  num_items,
  num.quad
)
}
\arguments{
\item{parm}{Item parameter being maximized.}

\item{p_item}{Vector of item parameters.}

\item{etable}{E-table for impact.}

\item{theta}{Matrix of adaptive theta values.}

\item{responses_item}{Vector of item responses.}

\item{pred.data}{Matrix or dataframe of DIF and/or impact predictors.}

\item{cov}{Covariate being maximized.}

\item{samp_size}{Sample size in dataset.}

\item{num_items}{Number of items in dataset.}

\item{num.quad}{Number of quadrature points used for approximating the
latent variable.}
}
\description{
Partial derivatives for variance parameter of continuous items.
}
\keyword{internal}
