% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derivatives.r
\name{d_sigma_gaussian_proxy}
\alias{d_sigma_gaussian_proxy}
\title{Partial derivatives for variance parameter of continuous items with proxy data.}
\usage{
d_sigma_gaussian_proxy(
  parm,
  p_item,
  prox_data,
  responses_item,
  pred_data,
  cov,
  samp_size,
  num_items
)
}
\arguments{
\item{parm}{Item parameter being maximized.}

\item{p_item}{Vector of item parameters.}

\item{prox_data}{Vector of observed proxy scores.}

\item{responses_item}{Vector of item responses.}

\item{pred_data}{Matrix or dataframe of DIF and/or impact predictors.}

\item{cov}{Covariate being maximized.}

\item{samp_size}{Sample size in dataset.}

\item{num_items}{Number of items in dataset.}
}
\value{
a \code{"list"} of first and second partial derivatives for variance value of Gaussian
item likelihood (to use with coordinate descent and univariate Newton-Raphson)
}
\description{
Partial derivatives for variance parameter of continuous items with proxy data.
}
\keyword{internal}
