% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_estimation.R
\name{em_estimation}
\alias{em_estimation}
\title{Penalized expectation-maximization algorithm.}
\usage{
em_estimation(
  p,
  item.data,
  predictor.data,
  mean_predictors,
  var_predictors,
  item.type,
  penalty.type,
  tau,
  alpha,
  gamma,
  pen,
  anchor,
  final.control,
  samp_size,
  num_items,
  num_responses,
  num_predictors,
  quadpts,
  adaptive.quad
)
}
\arguments{
\item{p}{List of parameters with starting values obtained from preprocess.}

\item{item.data}{Matrix or dataframe of item responses.}

\item{predictor.data}{Matrix or dataframe of DIF and/or impact predictors.}

\item{mean_predictors}{Possibly different matrix of predictors for the mean
impact equation.}

\item{var_predictors}{Possibly different matrix of predictors for the
variance impact equation.}

\item{item.type}{Optional character value or vector indicating the type of
item to be modeled.}

\item{penalty.type}{Character value indicating the penalty function to use.}

\item{tau}{Optional numeric vector of tau values.}

\item{alpha}{Numeric value indicating the alpha parameter in the elastic net
penalty function.}

\item{gamma}{Numeric value indicating the gamma parameter in the MCP
function.}

\item{anchor}{Optional numeric value or vector indicating which item
response(s) are anchors (e.g., \code{anchor = 1}).}

\item{quadpts}{Numeric value indicating the number of quadrature points.}

\item{ntau}{Numeric value of how many to tau values to fit.}

\item{tau.max}{Numberic value indicating the maximum tau parameter.}

\item{rasch}{Logical value indicating whether to constrain item slopes
to 1 (i.e., equal slopes).}

\item{impact.data}{Optional list of matrices or data frames with predictors
for mean and variance impact.}

\item{standardize}{Logical value indicating whether to standardize DIF and
impact covariates for regularization.}

\item{control}{Optional list of optimization parameters.}

\item{call}{Defined from regDIF.}
}
\description{
Penalized expectation-maximization algorithm.
}
\keyword{internal}
