% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_step.r
\name{Estep}
\alias{Estep}
\title{Expectation step.}
\usage{
Estep(
  p,
  item.data,
  predictor.data,
  mean_predictors,
  var_predictors,
  samp_size,
  num_items,
  num_responses,
  adaptive.quad,
  num_quadpts
)
}
\arguments{
\item{p}{List of parameters.}

\item{item.data}{Matrix or dataframe of item responses.}

\item{predictor.data}{Matrix or dataframe of DIF and/or impact predictors.}

\item{mean_predictors}{Possibly different matrix of predictors for the mean
impact equation.}

\item{var_predictors}{Possibly different matrix of predictors for the
variance impact equation.}

\item{samp_size}{Sample size in dataset.}

\item{num_items}{Number of items in dataset.}

\item{num_responses}{Number of responses for each item.}

\item{num_quadpts}{Number of quadrature points used for approximating the
latent variable.}
}
\description{
Expectation step.
}
\keyword{internal}
