% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traceline.r
\name{gaussian_traceline_pts}
\alias{gaussian_traceline_pts}
\title{Continuous tracelines.}
\usage{
gaussian_traceline_pts(
  p_active,
  theta,
  responses_item,
  pred.data,
  samp_size,
  quad.pts
)
}
\arguments{
\item{p_active}{Vector of item parameters.}

\item{theta}{Matrix of adaptive theta values.}

\item{responses_item}{Vector of item responses.}

\item{pred.data}{Matrix or dataframe of DIF and/or impact predictors.}

\item{samp_size}{Sample size in dataset.}

\item{quad.pts}{Number of quadrature points used for approximating the
latent variable.}
}
\description{
Continuous tracelines.
}
\keyword{internal}
