% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traceline.r
\name{gaussian_traceline_pts_proxy}
\alias{gaussian_traceline_pts_proxy}
\title{Continuous tracelines using proxy data.}
\usage{
gaussian_traceline_pts_proxy(
  p_item,
  prox_data,
  responses_item,
  pred_data,
  samp_size
)
}
\arguments{
\item{p_item}{Vector of item parameters.}

\item{prox_data}{Vector of observed proxy scores.}

\item{responses_item}{Vector of item responses.}

\item{pred_data}{Matrix or data frame of DIF and/or impact predictors.}

\item{samp_size}{Sample size in data set.}
}
\value{
a \code{"matrix"} of probability values for Gaussian item likelihood
}
\description{
Continuous tracelines using proxy data.
}
\keyword{internal}
