% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ida.R
\docType{data}
\name{ida}
\alias{ida}
\title{Simulated data example with multiple DIF covariates}
\format{
A data frame with 500 rows and 9 variables:
\describe{
  \item{item1}{}
  \item{item2}{}
  \item{item3}{}
  \item{item4}{}
  \item{item5}{}
  \item{item6}{}
  \item{age}{}
  \item{gender}{}
  \item{study}{}

  ...
}
}
\usage{
ida
}
\description{
A simulated dataset containing six binary items and three DIF covariates
}
\keyword{datasets}
