% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_tau.R
\name{identify_tau}
\alias{identify_tau}
\title{Identify the minimum value of tau such that all DIF is removed.}
\usage{
identify_tau(
  p,
  item_data,
  pred_data,
  mean_predictors,
  var_predictors,
  item_type,
  theta,
  pen_type,
  tau_vec,
  num_tau,
  alpha,
  gamma,
  anchor,
  final_control,
  samp_size,
  num_items,
  num_responses,
  num_predictors,
  num_quad,
  adapt_quad,
  optim_method,
  em_history
)
}
\arguments{
\item{p}{List of parameters with starting values obtained from preprocess.}

\item{item_data}{Matrix or dataframe of item responses.}

\item{pred_data}{Matrix or dataframe of DIF and/or impact predictors.}

\item{mean_predictors}{Possibly different matrix of predictors for the mean
impact equation.}

\item{var_predictors}{Possibly different matrix of predictors for the
variance impact equation.}

\item{item_type}{Optional character value or vector indicating the type of
item to be modeled.}

\item{theta}{Vector of fixed weights to approximate the latent variable.}

\item{pen_type}{Character value indicating the penalty function to use.}

\item{tau_vec}{Vector of tau values that either are automatically generated
or provided by the user.}

\item{num_tau}{Number of tau values to run Reg-DIF on.}

\item{alpha}{Numeric value indicating the alpha parameter in the elastic net
penalty function.}

\item{gamma}{Numeric value indicating the gamma parameter in the MCP
function.}

\item{anchor}{Optional numeric value or vector indicating which item
response(s) are anchors (e.g., \code{anchor = 1}).}

\item{final_control}{List of control parameters.}

\item{samp_size}{Numeric value indicating the sample size.}

\item{num_items}{Numeric value indicating the number of items.}

\item{num_responses}{Vector with number of responses for each item.}

\item{num_predictors}{Numeric value indicating the number of predictors.}

\item{num_quad}{Numeric value indicating the number of quadrature points.}

\item{adapt_quad}{Logical value indicating whether to use adaptive quad.}

\item{optim_method}{Character value indicating the type of optimization
method to use.}

\item{em_history}{List to save EM iterations for supplemental EM algorithm.}
}
\description{
Identify the minimum value of tau such that all DIF is removed.
}
\keyword{internal}
