% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infocriteria.R
\name{information_criteria}
\alias{information_criteria}
\title{Maximization step.}
\usage{
information_criteria(
  etable,
  p,
  item_data,
  pred_data,
  mean_predictors,
  var_predictors,
  gamma,
  samp_size,
  num_responses,
  num_items,
  num_quad
)
}
\arguments{
\item{etable}{Matrix of E-table values for item and impact equations.}

\item{p}{List of parameters.}

\item{item_data}{Matrix or dataframe of item responses.}

\item{pred_data}{Matrix or dataframe of DIF and/or impact predictors.}

\item{mean_predictors}{Possibly different matrix of predictors for the mean
impact equation.}

\item{var_predictors}{Possibly different matrix of predictors for the
variance impact equation.}

\item{gamma}{Numeric value indicating the gamma parameter in the MCP
function.}

\item{samp_size}{Sample size in data set.}

\item{num_responses}{Number of responses for each item.}

\item{num_items}{Number of items in data set.}

\item{num_quad}{Number of quadrature points used for approximating the
latent variable.}
}
\description{
Maximization step.
}
\keyword{internal}
