% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_step.r
\name{Mstep}
\alias{Mstep}
\title{Maximization step.}
\usage{
Mstep(
  p,
  item.data,
  predictor.data,
  mean_predictors,
  var_predictors,
  elist,
  item.type,
  penalty.type,
  tau,
  alpha,
  gamma,
  anchor,
  samp_size,
  num_responses,
  num_items,
  num_quadpts,
  num_predictors
)
}
\arguments{
\item{p}{List of parameters.}

\item{item.data}{Matrix or dataframe of item responses.}

\item{predictor.data}{Matrix or dataframe of DIF and/or impact predictors.}

\item{mean_predictors}{Possibly different matrix of predictors for the mean
impact equation.}

\item{var_predictors}{Possibly different matrix of predictors for the
variance impact equation.}

\item{elist}{List of E-tables for item and impact equations, in addition to
theta values.}

\item{item.type}{Optional character value or vector indicating the type of
item to be modeled.}

\item{penalty.type}{Character value indicating the penalty function to use.}

\item{tau}{Optional numeric vector of tau values.}

\item{alpha}{Numeric value indicating the alpha parameter in the elastic net
penalty function.}

\item{gamma}{Numeric value indicating the gamma parameter in the MCP
function.}

\item{anchor}{Optional numeric value or vector indicating which item
response(s) are anchors (e.g., \code{anchor = 1}).}

\item{samp_size}{Sample size in dataset.}

\item{num_responses}{Number of responses for each item.}

\item{num_items}{Number of items in dataset.}

\item{num_quadpts}{Number of quadrature points used for approximating the
latent variable.}

\item{num_predictors}{Number of predictors.}
}
\description{
Maximization step.
}
\keyword{internal}
