% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_tau.R
\name{Mstep_block_idtau}
\alias{Mstep_block_idtau}
\title{Final M-step using block estimation for identifying the minimum value of tau.}
\usage{
Mstep_block_idtau(
  p,
  item_data,
  pred_data,
  mean_predictors,
  var_predictors,
  etable,
  item_type,
  pen_type,
  tau_current,
  alpha,
  gamma,
  anchor,
  samp_size,
  num_responses,
  num_items,
  num_quad,
  num_predictors
)
}
\arguments{
\item{p}{List of parameters.}

\item{item_data}{Matrix or data frame of item responses.}

\item{pred_data}{Matrix or data frame of DIF and/or impact predictors.}

\item{mean_predictors}{Possibly different matrix of predictors for the mean
impact equation.}

\item{var_predictors}{Possibly different matrix of predictors for the
variance impact equation.}

\item{etable}{Etable for item and impact equations, in addition to
theta values.}

\item{item_type}{Optional character value or vector indicating the type of
item to be modeled.}

\item{pen_type}{Character value indicating the penalty function to use.}

\item{tau_current}{A single numeric value of tau that exists within
\code{tau_vec}.}

\item{alpha}{Numeric value indicating the alpha parameter in the elastic net
penalty function.}

\item{gamma}{Numeric value indicating the gamma parameter in the MCP
function.}

\item{anchor}{Optional numeric value or vector indicating which item
response(s) are anchors (e.g., \code{anchor = 1}).}

\item{samp_size}{Sample size in data set.}

\item{num_responses}{Number of responses for each item.}

\item{num_items}{Number of items in data set.}

\item{num_quad}{Number of quadrature points used for approximating the
latent variable.}

\item{num_predictors}{Number of predictors.}
}
\description{
Final M-step using block estimation for identifying the minimum value of tau.
}
\keyword{internal}
