% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.regdif.R
\name{plot.regDIF}
\alias{plot.regDIF}
\title{Plot function for regDIF function}
\usage{
\method{plot}{regDIF}(x, y, ...)
}
\arguments{
\item{x}{Fitted regDIF model object.}

\item{y}{Unused.}
}
\description{
Plot function for regDIF function
}
