% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{preprocess}
\alias{preprocess}
\title{Pre-process data.}
\usage{
preprocess(
  item.data,
  pred.data,
  item.type,
  num.tau,
  tau,
  anchor,
  stdz,
  num.quad,
  control,
  pen.type,
  adapt.quad,
  call
)
}
\arguments{
\item{item.data}{Matrix or data frame of item responses.}

\item{pred.data}{Matrix or data frame of DIF and/or impact predictors.}

\item{item.type}{Character value or vector indicating the item response
distributions.}

\item{num.tau}{Numeric value of how many to tau values to fit.}

\item{tau}{Optional numeric vector of tau values.}

\item{anchor}{Optional numeric value or vector indicating which item
response(s) are anchors (e.g., \code{anchor = 1}).}

\item{stdz}{Logical value indicating whether to standardize DIF and
impact predictors for regularization.}

\item{num.quad}{Numeric value indicating the number of quadrature points.}

\item{control}{Optional list of additional model specification and
optimization parameters.}

\item{pen.type}{Character indicating type of penalty.}

\item{adapt.quad}{Logical determining whether to use adaptive quadrature.}

\item{call}{Defined from regDIF.}
}
\description{
Pre-process data.
}
\keyword{internal}
