% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regDIF-package.R
\docType{package}
\name{regDIF-package}
\alias{regDIF-package}
\title{Regularized differential item functioning for IRT and CFA models.}
\description{
Regularized Differential Item Functioning
}
\details{
\pkg{regDIF} is a package that performs regularization of differential item functioning (DIF)
in item response theory (IRT) and confirmatory factor analysis (CFA) models using a
penalized expectation-maximization algorithm.
}
\author{
William Belzak \email{wbelzak@gmail.com}
}
\keyword{package}
