% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se.regDIF.R
\name{se.regDIF}
\alias{se.regDIF}
\title{Standard Errors for regDIF Model(s)}
\usage{
se.regDIF(model.obj,
          se.type = "sem",
          item.data = NULL,
          pred.data = NULL,
          tau = NULL,
          ...)
}
\arguments{
\item{model.obj}{A regDIF fitted model object of class \code{regDIF}.
Upon designating "model.obj", the default is to obtain standard errors for
the best-fitting model according to the minimum BIC model.}

\item{se.type}{Character value indicating the method of computing standard
errors for a regDIF fitted model. Default is "sem", or the supplemental EM
algorithm. Other options are not currently supported.}

\item{item.data}{Optional data frame or matrix of item response data to fit
regDIF model(s) and obtain standard errors. Default is NULL when
\code{model.obj} is provided. If providing \code{item.data}, then
\code{pred.data} and \code{tau} must also be provided.}

\item{pred.data}{Optional data frame or matrix of predictor data to fit
regDIF model(s) and obtain standard errors. Default is NULL when
\code{model.obj} is provided. If providing \code{pred.data}, then
\code{item.data} and \code{tau} must also be provided.}

\item{tau}{Optional numeric or vector of tau values to fit regDIF model(s)
and obtain standard errors. If providing \code{tau}, then either a
\code{model.obj} with a corresponding \code{tau} value (i.e., already fitted
to \code{tau}) must be provided or \code{item.data} and \code{pred.data}
must be provided.}

\item{...}{Additional arguments to pass to regDIF when \code{item.data},
\code{pred.data}, and \code{tau} are provided.}
}
\value{
Function returns an object of class \code{se.regDIF}
}
\description{
Obtain standard errors for regDIF model(s).
}
\examples{
\dontrun{

library(regDIF)
head(ida)
item.data <- ida[,1:6]
pred.data <- ida[,7:9]
fit <- regDIF(item.data, pred.data, tau = .2)
se.fit <- se.regDIF(fit)
se.fit

}

}
