% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.regDIF.R
\name{summary.regDIF}
\alias{summary.regDIF}
\title{Summary function for regDIF function}
\usage{
\method{summary}{regDIF}(object, ...)
}
\arguments{
\item{object}{Fitted regDIF model object.}

\item{...}{Additional arguments to be passed through \code{summary}.}
}
\value{
\code{NULL}
}
\description{
Summary function for regDIF function
}
