% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traceline.r
\name{bernoulli_traceline_pts}
\alias{bernoulli_traceline_pts}
\title{Binary item tracelines.}
\usage{
bernoulli_traceline_pts(p_item, theta, pred_data, samp_size)
}
\arguments{
\item{p_item}{Vector of item parameters.}

\item{theta}{Vector of theta values.}

\item{pred_data}{Matrix or dataframe of DIF and/or impact predictors.}

\item{samp_size}{Sample size in dataset.}
}
\value{
a \code{"matrix"} of probability values for Bernoulli item likelihood
}
\description{
Binary item tracelines.
}
\keyword{internal}
