% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.regDIF.R
\name{coef.regDIF}
\alias{coef.regDIF}
\title{Coefficient function for regDIF function}
\usage{
\method{coef}{regDIF}(object, lambda = NULL, method = "bic", ...)
}
\arguments{
\item{object}{Fitted regDIF model object.}

\item{lambda}{Optional character or numeric indicating the lambda(s) at which the model coefficients are returned. For character value, may be \code{"lambda.min"}, which returns model coefficients for the value of lambda at which the minimum fit statistic is identified. For numeric, the value(s) provided corresponds to the value(s) of lambda.}

\item{method}{Character value indicating the model fit statistic to be used for determining \code{"lambda.min"}. Default is \code{"bic"}. May also be \code{"aic"}.}

\item{...}{Additional arguments to be passed through.}
}
\description{
Coefficient function for regDIF function
}
