% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.regDIF.R
\name{coef.regDIF}
\alias{coef.regDIF}
\title{Coefficient function for regDIF function}
\usage{
\method{coef}{regDIF}(x)
}
\arguments{
\item{x}{regDIF model object to obtain coefficient values.}
}
\value{
\code{NULL}
}
\description{
Coefficient function for regDIF function
}
