% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.regDIF.R
\name{coef.regDIF}
\alias{coef.regDIF}
\title{Coefficient function for regDIF function}
\usage{
\method{coef}{regDIF}(object, ...)
}
\arguments{
\item{object}{regDIF model object to obtain coefficient values.}

\item{...}{Additional arguments to be passed through}
}
\value{
\code{NULL}
}
\description{
Coefficient function for regDIF function
}
