% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traceline.r
\name{cumulative_traceline_pts}
\alias{cumulative_traceline_pts}
\title{Ordinal tracelines (for derivatives).}
\usage{
cumulative_traceline_pts(
  p_item,
  theta,
  pred_data,
  samp_size,
  num_responses_item,
  num_quad
)
}
\arguments{
\item{p_item}{Vector of item parameters.}

\item{theta}{Vector of theta values.}

\item{pred_data}{Matrix or dataframe of DIF and/or impact predictors.}

\item{samp_size}{Sample size in dataset.}

\item{num_responses_item}{Number of responses for item.}

\item{num_quad}{Number of quadrature points used for approximating the
latent variable.}
}
\value{
a \code{"matrix"} of probability values for categorical (cumulative) item likelihood
}
\description{
Ordinal tracelines (for derivatives).
}
\keyword{internal}
