% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derivatives.r
\name{d_categorical_itemblock}
\alias{d_categorical_itemblock}
\title{Partial derivatives for ordinal items.}
\usage{
d_categorical_itemblock(
  parm,
  p_item,
  etable,
  theta,
  pred_data,
  item_data_current,
  samp_size,
  num_responses_item,
  num_items,
  num_predictors,
  num_quad
)
}
\arguments{
\item{parm}{Item parameter being maximized.}

\item{p_item}{Vector of item parameters.}

\item{etable}{E-table for impact.}

\item{theta}{Matrix of adaptive theta values.}

\item{pred_data}{Matrix or dataframe of DIF and/or impact predictors.}

\item{item_data_current}{Vector of current item responses.}

\item{samp_size}{Sample size in dataset.}

\item{num_items}{Number of items in dataset.}

\item{num_predictors}{Number of predictors in dataset.}

\item{num_quad}{Number of quadrature points used for approximating the
latent variable.}
}
\value{
a \code{"list"} of first and second partial derivatives for categorical item likelihood
(to use with multivariate Newton-Raphson)
}
\description{
Partial derivatives for ordinal items.
}
\keyword{internal}
