% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derivatives.r
\name{d_phi}
\alias{d_phi}
\title{Partial derivatives for mean impact equation.}
\usage{
d_phi(
  p_impact,
  etable_all,
  theta,
  mean_predictors,
  var_predictors,
  cov,
  samp_size,
  num_items,
  num_quadpts
)
}
\arguments{
\item{etable_all}{E-table for impact.}

\item{theta}{Matrix of adaptive theta values.}

\item{mean_predictors}{Possibly different matrix of predictors for the mean
impact equation.}

\item{var_predictors}{Possibly different matrix of predictors for the
variance impact equation.}

\item{cov}{Covariate being maximized.}

\item{samp_size}{Sample size in dataset.}

\item{num_items}{Number of items in dataset.}

\item{num_quadpts}{Number of quadrature points used for approximating the
latent variable.}

\item{p_active}{Vector of item parameters.}
}
\description{
Partial derivatives for mean impact equation.
}
\keyword{internal}
