% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derivatives.r
\name{d_sigma_gaussian}
\alias{d_sigma_gaussian}
\title{Partial derivatives for variance parameter of continuous items.}
\usage{
d_sigma_gaussian(
  parm,
  p_item,
  etable_item,
  theta,
  responses_item,
  pred_data,
  cov,
  samp_size,
  num_items,
  num_quad
)
}
\arguments{
\item{parm}{Item parameter being maximized.}

\item{p_item}{Vector of item parameters.}

\item{etable_item}{E-table for impact.}

\item{theta}{Matrix of adaptive theta values.}

\item{responses_item}{Vector of item responses.}

\item{pred_data}{Matrix or dataframe of DIF and/or impact predictors.}

\item{cov}{Covariate being maximized.}

\item{samp_size}{Sample size in dataset.}

\item{num_items}{Number of items in dataset.}

\item{num_quad}{Number of quadrature points used for approximating the
latent variable.}
}
\value{
a \code{"list"} of first and second partial derivatives for variance value of Gaussian
item likelihood (to use with coordinate descent and univariate Newton-Raphson)
}
\description{
Partial derivatives for variance parameter of continuous items.
}
\keyword{internal}
