% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/em_estimation.R
\name{em_estimation}
\alias{em_estimation}
\title{Penalized expectation-maximization algorithm.}
\usage{
em_estimation(
  p,
  item.data,
  pred.data,
  mean_predictors,
  var_predictors,
  item.type,
  pen.type,
  tau_vec,
  alpha,
  gamma,
  pen,
  anchor,
  final.control,
  samp_size,
  num_items,
  num_responses,
  num_predictors,
  num.quad,
  adapt.quad
)
}
\arguments{
\item{p}{List of parameters with starting values obtained from preprocess.}

\item{item.data}{Matrix or data frame of item responses.}

\item{pred.data}{Matrix or data frame of DIF and/or impact predictors.}

\item{mean_predictors}{Possibly different matrix of predictors for the mean
impact equation.}

\item{var_predictors}{Possibly different matrix of predictors for the
variance impact equation.}

\item{item.type}{Optional character value or vector indicating the type of
item to be modeled.}

\item{pen.type}{Character value indicating the penalty function to use.}

\item{tau_vec}{Vector of tau values that either are automatically generated
or provided by the user. The first \code{tau_vec} will be equal to \code{Inf}
to identify a minimal value of tau in which all DIF is removed from the
model.}

\item{alpha}{Numeric value indicating the alpha parameter in the elastic net
penalty function.}

\item{gamma}{Numeric value indicating the gamma parameter in the MCP
function.}

\item{pen}{Index for the tau vector.}

\item{anchor}{Optional numeric value or vector indicating which item
response(s) are anchors (e.g., \code{anchor = 1}).}

\item{final.control}{Control parameters.}

\item{samp_size}{Numeric value indicating the sample size.}

\item{num_items}{Numeric value indicating the number of items.}

\item{num_responses}{Vector with number of responses for each item.}

\item{num_predictors}{Numeric value indicating the number of predictors.}

\item{num.quad}{Numeric value indicating the number of quadrature points.}

\item{adapt.quad}{Logical value indicating whether to use adaptive quad.
needs to be identified.}
}
\description{
Penalized expectation-maximization algorithm.
}
\keyword{internal}
