% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traceline.r
\name{gaussian_traceline_pts}
\alias{gaussian_traceline_pts}
\title{Continuous traceline values.}
\usage{
gaussian_traceline_pts(
  p_active,
  theta,
  responses_item,
  predictor.data,
  samp_size,
  num_quadpts
)
}
\arguments{
\item{p_active}{Vector of item parameters.}

\item{theta}{Matrix of adaptive theta values.}

\item{responses_item}{Vector of item responses.}

\item{predictor.data}{Matrix or dataframe of DIF and/or impact predictors.}

\item{samp_size}{Sample size in dataset.}

\item{num_quadpts}{Number of quadrature points used for approximating the
latent variable.}
}
\description{
Continuous traceline values.
}
\keyword{internal}
