% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traceline.r
\name{gaussian_traceline_pts}
\alias{gaussian_traceline_pts}
\title{Continuous tracelines.}
\usage{
gaussian_traceline_pts(p_active, theta, responses_item, pred.data, samp_size)
}
\arguments{
\item{p_active}{Vector of item parameters.}

\item{theta}{Vector of theta values.}

\item{responses_item}{Vector of item responses.}

\item{pred.data}{Matrix or dataframe of DIF and/or impact predictors.}

\item{samp_size}{Sample size in dataset.}
}
\description{
Continuous tracelines.
}
\keyword{internal}
