% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traceline.r
\name{gaussian_traceline_pts}
\alias{gaussian_traceline_pts}
\title{Continuous tracelines.}
\usage{
gaussian_traceline_pts(p_item, theta, responses_item, pred_data, samp_size)
}
\arguments{
\item{p_item}{Vector of item parameters.}

\item{theta}{Vector of theta values.}

\item{responses_item}{Vector of item responses.}

\item{pred_data}{Matrix or data frame of DIF and/or impact predictors.}

\item{samp_size}{Sample size in data set.}
}
\value{
a \code{"matrix"} of probability values for Gaussian item likelihood
}
\description{
Continuous tracelines.
}
\keyword{internal}
