% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infocriteria.R
\name{information_criteria}
\alias{information_criteria}
\title{Maximization step.}
\usage{
information_criteria(
  elist,
  p,
  item.data,
  pred.data,
  mean_predictors,
  var_predictors,
  theta,
  gamma,
  samp_size,
  num_responses,
  num_items,
  quad.pts
)
}
\arguments{
\item{elist}{List of E-tables for item and impact equations, in addition to
theta values.}

\item{p}{List of parameters.}

\item{item.data}{Matrix or dataframe of item responses.}

\item{pred.data}{Matrix or dataframe of DIF and/or impact predictors.}

\item{mean_predictors}{Possibly different matrix of predictors for the mean
impact equation.}

\item{var_predictors}{Possibly different matrix of predictors for the
variance impact equation.}

\item{theta}{Matrix of adaptive theta values.}

\item{gamma}{Numeric value indicating the gamma parameter in the MCP
function.}

\item{samp_size}{Sample size in dataset.}

\item{num_responses}{Number of responses for each item.}

\item{num_items}{Number of items in dataset.}

\item{quad.pts}{Number of quadrature points used for approximating the
latent variable.}
}
\description{
Maximization step.
}
\keyword{internal}
