% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infocriteria.R
\name{information_criteria}
\alias{information_criteria}
\title{Maximization step.}
\usage{
information_criteria(
  eout,
  p,
  item_data,
  pred_data,
  prox_data,
  mean_predictors,
  var_predictors,
  gamma,
  samp_size,
  num_responses,
  num_items,
  num_quad
)
}
\arguments{
\item{eout}{E-table output.}

\item{p}{List of parameters.}

\item{item_data}{Matrix or data.frame of item responses.}

\item{pred_data}{Matrix or data.frame of DIF and/or impact predictors.}

\item{prox_data}{Vector of observed proxy scores.}

\item{mean_predictors}{Possibly different matrix of predictors for the mean
impact equation.}

\item{var_predictors}{Possibly different matrix of predictors for the
variance impact equation.}

\item{gamma}{Numeric value indicating the gamma parameter in the MCP
function.}

\item{samp_size}{Sample size in data set.}

\item{num_responses}{Number of responses for each item.}

\item{num_items}{Number of items in data set.}

\item{num_quad}{Number of quadrature points used for approximating the
latent variable.}
}
\value{
a \code{"list"} of information criteria to use for model selection
}
\description{
Maximization step.
}
\keyword{internal}
