% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_step.r
\name{Mstep}
\alias{Mstep}
\title{Maximization step.}
\usage{
Mstep(
  p,
  item.data,
  pred.data,
  mean_predictors,
  var_predictors,
  etable,
  theta,
  item.type,
  pen.type,
  tau_current,
  alpha,
  gamma,
  anchor,
  samp_size,
  num_responses,
  num_items,
  num.quad,
  num_predictors
)
}
\arguments{
\item{p}{List of parameters.}

\item{item.data}{Matrix or data frame of item responses.}

\item{pred.data}{Matrix or data frame of DIF and/or impact predictors.}

\item{mean_predictors}{Possibly different matrix of predictors for the mean
impact equation.}

\item{var_predictors}{Possibly different matrix of predictors for the
variance impact equation.}

\item{etable}{E-table matrix for item and impact equations.}

\item{theta}{Vector of fixed quadrature points.}

\item{item.type}{Optional character value or vector indicating the type of
item to be modeled.}

\item{pen.type}{Character value indicating the penalty function to use.}

\item{tau_current}{A single numeric value of tau that exists within
\code{tau_vec}.}

\item{alpha}{Numeric value indicating the alpha parameter in the elastic net
penalty function.}

\item{gamma}{Numeric value indicating the gamma parameter in the MCP
function.}

\item{anchor}{Optional numeric value or vector indicating which item
response(s) are anchors (e.g., \code{anchor = 1}).}

\item{samp_size}{Sample size in data set.}

\item{num_responses}{Number of responses for each item.}

\item{num_items}{Number of items in data set.}

\item{num.quad}{Number of quadrature points used for approximating the
latent variable.}

\item{num_predictors}{Number of predictors.}
}
\description{
Maximization step.
}
\keyword{internal}
