% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.regdif.R
\name{plot.regDIF}
\alias{plot.regDIF}
\title{Plot function for regDIF function}
\usage{
\method{plot}{regDIF}(x, y = NULL, method = "bic", color.seed = 123, legend.plot = TRUE, ...)
}
\arguments{
\item{x}{Fitted regDIF model object.}

\item{y}{Unused for plotting regDIF model object.}

\item{method}{Fit statistic to use for identifying DIF effects in plot.}

\item{color.seed}{Random seed to sample line colors and line types for
DIF effects in plot.}

\item{legend.plot}{Logical indicating whether to plot a legend. Default is \code{TRUE}.}

\item{...}{Additional arguments to be passed through to \code{plot}.}
}
\value{
a \code{"plot"} object for a \code{"regDIF"} fit
}
\description{
Plot function for regDIF function
}
