% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess.R
\name{postprocess}
\alias{postprocess}
\title{Maximization step.}
\usage{
postprocess(
  estimates,
  item.data,
  pred.data,
  mean_predictors,
  var_predictors,
  tau_vec,
  alpha,
  pen,
  anchor,
  control,
  final.control,
  final,
  samp_size,
  num_responses,
  num_predictors,
  num_items,
  num.quad
)
}
\arguments{
\item{estimates}{List of converged parameters.}

\item{item.data}{Matrix or dataframe of item responses.}

\item{pred.data}{Matrix or dataframe of DIF and/or impact predictors.}

\item{mean_predictors}{Possibly different matrix of predictors for the mean
impact equation.}

\item{var_predictors}{Possibly different matrix of predictors for the
variance impact equation.}

\item{tau_vec}{Optional numeric vector of tau values.}

\item{alpha}{Numeric value indicating the alpha parameter in the elastic net
penalty function.}

\item{pen}{Tuning parameter index.}

\item{anchor}{Anchor item(s).}

\item{control}{Optional list of user-defined control parameters}

\item{final.control}{List of final control parameters.}

\item{final}{List of model results.}

\item{samp_size}{Sample size in dataset.}

\item{num_responses}{Number of responses for each item.}

\item{num_predictors}{Number of predictors.}

\item{num_items}{Number of items in dataset.}

\item{num_quadpts}{Number of quadrature points used for approximating the
latent variable.}
}
\description{
Maximization step.
}
\keyword{internal}
