% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{preprocess}
\alias{preprocess}
\title{Pre-process data.}
\usage{
preprocess(
  item.data,
  predictor.data,
  item.type,
  num.tau,
  max.tau,
  tau,
  anchor,
  rasch,
  impact.data,
  standardize,
  quadpts,
  control,
  call
)
}
\arguments{
\item{item.data}{Matrix or dataframe of item responses.}

\item{item.type}{Character value or vector indicating the item response
distributions.}

\item{num.tau}{Numeric value of how many to tau values to fit.}

\item{max.tau}{Numberic value indicating the maximum tau parameter.}

\item{tau}{Optional numeric vector of tau values.}

\item{anchor}{Optional numeric value or vector indicating which item
response(s) are anchors (e.g., \code{anchor = 1}).}

\item{rasch}{Logical value indicating whether to constrain item slopes
to 1 (i.e., equal slopes).}

\item{impact.data}{Optional list of matrices or data frames with predictors
for mean and variance impact.}

\item{standardize}{Logical value indicating whether to standardize DIF and
impact covariates for regularization.}

\item{quadpts}{Numeric value indicating the number of quadrature points.}

\item{control}{Optional list of optimization parameters.}

\item{call}{Defined from regDIF.}

\item{Matrix}{or dataframe of DIF and/or impact predictors.}
}
\description{
Pre-process data.
}
\keyword{internal}
