% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{preprocess}
\alias{preprocess}
\title{Pre-process data.}
\usage{
preprocess(
  item.data,
  pred.data,
  prox.data,
  item.type,
  pen.type,
  tau,
  num.tau,
  anchor,
  stdz,
  control,
  call
)
}
\arguments{
\item{item.data}{Matrix or data frame of item responses.}

\item{pred.data}{Matrix or data frame of DIF and/or impact predictors.}

\item{prox.data}{Vector of observed proxy scores.}

\item{item.type}{Character value or vector indicating the item response
distributions.}

\item{pen.type}{Character indicating type of penalty.}

\item{tau}{Optional numeric vector of tau values.}

\item{num.tau}{Numeric indicating number of tau values to run Reg-DIF on.}

\item{anchor}{Optional numeric value or vector indicating which item
response(s) are anchors (e.g., \code{anchor = 1}).}

\item{stdz}{Logical value indicating whether to standardize DIF and
impact predictors for regularization.}

\item{control}{Optional list of additional model specification and
optimization parameters.}

\item{call}{Defined from regDIF.}
}
\value{
a \code{"list"} of default controls for \code{"em_estimation"}
}
\description{
Pre-process data.
}
\keyword{internal}
