% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.regdif.R
\name{print.regDIF}
\alias{print.regDIF}
\title{Print function for regDIF function}
\usage{
\method{print}{regDIF}(x)
}
\arguments{
\item{x}{regDIF model object to obtain basic results}
}
\value{
\code{NULL}
}
\description{
Print function for regDIF function
}
