% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regdif.R
\name{regDIF}
\alias{regDIF}
\title{Regularized Differential Item Functioning}
\usage{
regDIF(data, covariates, tau, anchor = NULL, quadpts = 15,
  standardize = TRUE)
}
\arguments{
\item{data}{A matrix or data frame of item responses. Currently supports dichotomously (0-1) scored items only.}

\item{covariates}{A matrix or data frame of DIF covariates. Supports both categorical and continuous covariates.}

\item{tau}{A numeric vector of tuning parameter values. Default is \code{0} or no penalty.}

\item{anchor}{Optional. A numeric vector indicating which items are anchors. Default is \code{NULL}, which means that at least one DIF parameter per covariate across all items will be fixed to zero to identify the model.}

\item{quadpts}{The number of quadrature points to approximate the latent variable. More points lead to more precise estimates but slower run time. Default is \code{15} quadrature points.}

\item{standardize}{Logical (\code{TRUE}/\code{FALSE}) indicating whether to normalize the covariates. Default is \code{TRUE}.}
}
\value{
Function returns an object of class \code{SingleGroupClass}
}
\description{
Regularization of DIF parameters in item response theory (IRT) and moderated nonlinear factor analysis (MNLFA) models.
}
\examples{
\dontrun{

library(regDIF)
data <- ida[,1:6]
dif.covariates <- ida[,7:9]
tau <- seq(2,0,-.05)
model <- regDIF(data, dif.covariates, tau)
model

}

}
