% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regdif.R
\name{regDIF}
\alias{regDIF}
\title{Regularized Differential Item Functioning}
\usage{
regDIF(x, y, itemtypes = c("bernoulli", "categorical", "gaussian"),
  penalty = c("lasso", "mcp"), nlambda = 100, lambda.max = 3,
  gamma = 3, lambda = NULL, anchor = NULL, rasch = FALSE,
  control = list())
}
\arguments{
\item{x}{Matrix or dataframe of predictors (i.e., DIF covariates). Supports categorical and continuous predictors.}

\item{y}{Matrix or dataframe of item responses. Supports Bernoulli (e.g., 0,1), categorical (e.g., 0,1,2,...,k), and Gaussian item responses.}

\item{itemtypes}{Character value or vector indicating the item response distributions. For scales where item responses are of one type only, the user may input one character value indicating the type (e.g., \code{"categorical"}). For mixed item types, the user must specify a vector of characters in the order that corresponds to the response matrix (e.g. \code{c(rep("categorical",3),rep("gaussian",3))}). Supports categorical and continuous responses, including Bernoulli with logistic link (i.e., binary outcomes), categorical with ordered logistic link (i.e., graded response model), and Gaussian with identity link (i.e., factor analysis).}

\item{penalty}{Character value indicating the penalty function to use. Supports the least absolute selection and shrinkage operator (lasso) and the minimax concave penalty (mcp).}

\item{nlambda}{Numeric value indicating the number of lambda values to fit.}

\item{lambda.max}{Numberic value indicating the maximum lambda (shrinkage tuning parameter) to fit. Default is 3.}

\item{gamma}{Numeric value indicating the gamma (tapering tuning parameter) to fit. Default is 3.}

\item{lambda}{Optional numeric vector of tuning parameter values \eqn{\ge} 0. If supplied, must be in descending order, from largest to smallest values (e.g., \code{seq(1,0,-.01)}.}

\item{anchor}{Optional numeric vector indicating which items are anchors (e.g., \code{anchor = 1}). Default is \code{NULL}, meaning at least one DIF parameter (per covariate) will be fixed to zero as lambda approaches 0. This is required to identify the model.}

\item{rasch}{Logical value indicating whether to constrain item slopes to 1 (i.e., equal slopes). If \code{TRUE}, no slope DIF will be evaluated. Default is \code{FALSE}.}

\item{control}{Optional list of control parameters. See documentation.}
}
\value{
Function returns an object of class \code{regDIF}
}
\description{
Regularization of DIF parameters in item response theory (IRT) and moderated nonlinear factor analysis (MNLFA) models.
}
\examples{
\dontrun{

library(regDIF)
head(ida)
y <- ida[,1:6]
x <- ida[,7:9]
fit <- regDIF(x, y, itemtypes = "bernoulli", penalty = "lasso")
fit

}

}
