% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regDIF.R
\name{regDIF}
\alias{regDIF}
\title{Regularized Differential Item Functioning}
\usage{
regDIF(item.data,
       predictor.data,
       item.type = NULL,
       penalty.type = NULL,
       tau = NULL,
       num.tau = 100,
       max.tau = 2,
       alpha = 1,
       gamma = 3,
       anchor = NULL,
       impact.data = list(mean = NULL, var = NULL),
       standardize = TRUE,
       adaptive.quad = TRUE,
       quadpts = 15,
       control = list())
}
\arguments{
\item{item.data}{Matrix or dataframe of item responses. See below for
supported item types.}

\item{predictor.data}{Matrix or dataframe of DIF and/or impact predictors.
See below for option to specify different predictors for impact model.}

\item{item.type}{Optional character value or vector indicating the type of
item to be modeled. The default is NULL, corresponding to a 2PL or graded
item type. Different item types may be specified for a single model
by providing a vector equal in length to the number of items in item.data.
The options include:
\itemize{
   \item{\code{"rasch"} - Slopes constrained to 1 and intercepts freely
   estimated.}
   \item{\code{"2pl"} - Slopes and intercepts freely estimated.}
   \item{\code{"graded"} - Slopes, intercepts, and thresholds freely
   estimated.}
   \item{\code{"cfa"} - Not currently supported. Slopes and intercepts
   freely estimated with continuous (Gaussian) item responses.}}}

\item{penalty.type}{Optional character value indicating the penalty
function to use. The default is NULL, corresponding to the LASSO function.
The options include:
\itemize{
   \item{\code{"lasso"} - The least absolute selection and shrinkage
   operator (LASSO), which controls DIF selection through \eqn{\tau} (tau).}
   \item{\code{"mcp"} - The minimax concave penalty (MCP), which controls
   DIF selection through \eqn{\tau} (tau) and estimator bias through
   \eqn{\gamma} (gamma). Uses the firm-thresholding penalty function.}}}

\item{tau}{Optional numeric vector of tau values \eqn{\ge} 0. If tau is
supplied, this overrides the automatic construction of tau values via
\code{max.tau}. Must be non-negative and in descending order, from largest
to smallest values (e.g., \code{seq(1,0,-.01)}.}

\item{num.tau}{Numeric value indicating how many tau values to fit. The
default is 100.}

\item{max.tau}{Numberic value indicating the maximum tau parameter to use
for internal construction of tau vector. Default is 2. Must be large enough
to shrink all DIF effects to zero to begin with.}

\item{alpha}{Numeric value indicating the alpha parameter in the elastic net
penalty function. Alpha controls the degree to which LASSO or ridge is used
during regularization. The default is 1, which is equivalent to LASSO.
NOTE: If using MCP penalty, alpha may not be exactly 0.}

\item{gamma}{Numeric value indicating the gamma parameter in the MCP
function. Gamma controls the degree of tapering of DIF effects as tau
decreases. Larger gamma leads to faster tapering (less bias but possibly
more unstable optimization), whereas smaller gamma leads to slower tapering
(more bias but more stable optimization). Default is 3. Must be greater
than 1.}

\item{anchor}{Optional numeric value or vector indicating which item
response(s) are anchors (e.g., \code{anchor = 1}). Default is \code{NULL},
meaning at least one DIF effect per covariate will be fixed to zero as
tau approaches 0 (required to identify the model).}

\item{impact.data}{Optional list of matrices or data frames with predictors
for mean and variance impact. Allows for different sets of predictors on
the mean and variance impact equations compared to the item response DIF
equations.}

\item{standardize}{Logical value indicating whether to standardize DIF and
impact covariates for regularization. Default is \code{TRUE}, as it is
recommended that all covariates be on the same scale.}

\item{adaptive.quad}{Logical value indicating whether to use adaptive
quadrature. The default is \code{TRUE}. If set to \code{FALSE}, it is
recommended that 50 or more points be used.}

\item{quadpts}{Numeric value indicating the number of quadrature points to
be used in approximating the latent variable distribution during estimation.
Default is \code{15} when using adaptive quadrature.
be used in approximating the latent variable distribution during estimation.
Uses adaptive quadrature. Default is \code{15}.}

\item{control}{Optional list of optimization parameters. May be:
\describe{
   \item{tol}{Convergence threshold of EM algorithm. Default is
   \code{10^-5}.}
   \item{maxiter}{Maximum number of EM iterations. Default is
   \code{10000}.}}}
}
\value{
Function returns an object of class \code{regDIF}
}
\description{
Identify DIF in item response theory models using regularization.
}
\examples{
\dontrun{

library(regDIF)
head(ida)
item.data <- ida[,1:6]
predictor.data <- ida[,7:9]
fit <- regDIF(item.data, predictor.data)
summary(fit)

}

}
