% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regDIF.R
\name{regDIF}
\alias{regDIF}
\title{Regularized Differential Item Functioning}
\usage{
regDIF(data, covariates, tau, anchor = 1, quadpts = 15,
  standardize = TRUE)
}
\arguments{
\item{data}{A matrix or data frame. Currently supports dichotomously (0-1) scored items only.}

\item{covariates}{A matrix or data frame of DIF covariates. Supports both categorical and continuous covariates.}

\item{tau}{A numeric vector of tuning parameters.}

\item{anchor}{A number or numeric vector indicating which items are anchors. Currently at least one item must be specified as an anchor, although more anchors may be specified. Default is item \code{1}.}

\item{quadpts}{The number of quadrature points to approximate the latent variable. More points lead to more precise estimates but slower run time. Default is \code{15} quadrature points.}

\item{standardize}{A number indicating which covariate to standardize. It is recommended that all covariates are on the same scale or standardized (i.e., Normal(0,1)). Default is \code{0} (no covariates are standardized).}
}
\value{
Function returns an object of class \code{SingleGroupClass}
}
\description{
Regularization of DIF parameters in item response theory (IRT) and moderated nonlinear factor analysis (MNLFA) models.
}
\examples{
\dontrun{

library(regDIF)
data <- ida[,1:6]
dif.covariates <- ida[,7:9]
tau <- seq(1:100)
model <- regDIF(data, dif.covariates, tau)
model

}

}
