% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regdif.R
\name{regDIF}
\alias{regDIF}
\title{Regularized Differential Item Functioning}
\usage{
regDIF(data, covariates, penalty, standardize = TRUE, anchor = NULL,
  quadpts = 15)
}
\arguments{
\item{data}{Matrix or data frame of item responses. Currently supports dichotomously (0-1) scored items only.}

\item{covariates}{Matrix or data frame of DIF covariates. Supports both categorical and continuous covariates.}

\item{penalty}{Numeric vector of non-zero tuning parameter values.}

\item{standardize}{Logical value indicating whether to normalize the covariates. Default is \code{TRUE}.}

\item{anchor}{Optional numeric vector indicating which items are anchors (e.g., \code{anchor = 1}). Default is \code{NULL}, which means that at least one DIF parameter per covariate across all items will be fixed to zero to identify the model.}

\item{quadpts}{Number of quadrature points to approximate the latent variable. More points lead to more precise estimates but slower run time. Default is \code{15} quadrature points.}
}
\value{
Function returns an object of class \code{SingleGroupClass}
}
\description{
Regularization of DIF parameters in item response theory (IRT) and moderated nonlinear factor analysis (MNLFA) models.
}
\examples{
\dontrun{

library(regDIF)
data <- ida[,1:6]
covariates <- ida[,7:9]
penalty <- seq(1,0,-.05)
model <- regDIF(data, covariates, penalty)
model

}

}
